+++ backend/src/main/java/com/prograngers/backend/service/MemberService.java
@@ -1,5 +1,12 @@
-package com.prograngers.backend;
+package com.prograngers.backend.service;
 
+import com.prograngers.backend.AuthResult;
+import com.prograngers.backend.Encrypt;
+import com.prograngers.backend.JwtTokenProvider;
+import com.prograngers.backend.dto.LoginRequest;
+import com.prograngers.backend.RefreshToken;
+import com.prograngers.backend.repository.RefreshTokenRepository;
+import com.prograngers.backend.dto.SignUpRequest;
 import com.prograngers.backend.entity.Member;
 import com.prograngers.backend.exception.InvalidPasswordException;
 import com.prograngers.backend.exception.RefreshTokenNotFoundException;
